requirements.txt
# MIDAS Unified Framework Requirements

# Core Deep Learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
pytorch-lightning>=1.9.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
Pillow>=9.0.0

# NLP (for Food101 text processing)
transformers>=4.20.0

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.0

# Optional but recommended
tensorboard>=2.9.0
